Reorgan.  structure & network collection+config
