+++ src/collection.py
@@ -1,126 +1,12 @@
-"""This module contains all classes and functions for the OSM data collection
+"""This module contains all functions for the OSM data collection
 with the Pyrosm package."""
-import os
-import sys
 import time
 import yaml
 from pathlib import Path
-
-#from pygeos import GEOSException
 from pyrosm import get_data, OSM
-import geopandas as gpd
 import pandas as pd
 import numpy as np
-from src.config.osm_dict import OSM_tags
-class Config:
-    def __init__(self,name):
-        with open(Path(__file__).parent/'config/config.yaml', encoding="utf-8") as stream:
-            config = yaml.safe_load(stream)
-        var = config['VARIABLES_SET']
-        self.name = name
-        if list(var[name].keys()) == ['collection', 'preparation', 'fusion']:
-            self.pbf_data = var['region_pbf']
-            self.collection = var[name]['collection']
-            self.preparation = var[name]['preparation']
-            self.fusion = var[name]['fusion']
-        elif list(var[name].keys()) == ['variable_container']:
-            self.variable_container = var[name]['variable_container']
-        else:
-            print("unknown config format")
-            sys.exit()
-
-    def pyrosm_filter(self):
-        """creates a filter based on user input in the config to filter the OSM import"""
-        coll = self.collection
-
-        # check if input osm_tags, osm_features are valid and print non valid ones
-        for i in coll["osm_tags"].keys():
-            if i not in OSM_tags.keys():
-                print(f"{i} is not a valid osm_feature")
-        for i in [item for sublist in coll["osm_tags"].values() for item in sublist]:
-            if i not in [item for sublist in OSM_tags.values() for item in sublist] + ['all', True]:
-                print(f"{i} is not a valid osm_feature")
-
-        # loop collects all tags of a feature from osm_feature_tags_dict.py if "all" in config file
-        temp = {}
-        for key, values in coll["osm_tags"].items():
-            for value in values:
-                if value == 'all':
-                    temp = temp | {key:OSM_tags[key]}
-        
-        po_filter = coll["osm_tags"] | temp,None,"keep",list(coll["osm_tags"].keys())+\
-                    coll["additional_columns"],coll["points"], coll["lines"],coll["polygons"], None
-
-        return po_filter
-
-    def fusion_key_set(self, typ):
-        fus = self.fusion
-        try:
-            key_set = fus["fusion_data"]['source'][typ].keys()
-        except:
-            key_set = []
-        return key_set
-
-    def fusion_set(self,typ,key):
-        fus = self.fusion["fusion_data"]["source"][typ][key]
-        fus_set = fus["amenity"],fus["amenity_set"],fus["amenity_operator"],fus["columns2rename"], fus["column_set_value"], fus["columns2fuse"]
-        return fus_set
-    
-    def fusion_type(self, typ, key):
-        fus = self.fusion["fusion_data"]["source"][typ][key]
-        fus_type = fus["fusion_type"]
-        return fus_type
-
-def classify_osm_tags(name):
-    """helper function to help assign osm tags to their corresponding feature"""
-    # import dict from conf_yaml
-    with open(Path(__file__).parent/'config/config.yaml', encoding="utf-8") as stream:
-        config = yaml.safe_load(stream)
-    var = config['VARIABLES_SET']
-    temp = {}
-    for key in var[name]['collection']['osm_tags'].keys():
-        if key == 'not_sure':
-            for i in var[name]['collection']['osm_tags'][key]:
-                for keys, values in OSM_tags.items():
-                    for value in values:
-                        if i == value:
-                            if keys in temp.keys():
-                                if isinstance(temp[keys], str) is True:
-                                    temp[keys] = [temp[keys], i]
-                                else:
-                                    temp[keys].append(i)
-                            else:
-                                temp = temp | {keys:i}
-                        elif "no_valid_osm_tag" not in temp.keys() and i not in temp.values():
-                            temp = temp | {"no_valid_osm_tag":i}
-                        elif "no_valid_osm_tag" in temp.keys() and i not in temp["no_valid_osm_tag"]:
-                            if isinstance(temp["no_valid_osm_tag"], str) is True:
-                                temp["no_valid_osm_tag"] = [temp["no_valid_osm_tag"], i]
-                            else:
-                                temp["no_valid_osm_tag"].append(i)
-            print(temp)
-            sys.exit()
-
-# Function for creation backupfiles
-# Use it as return result for data preparation functions
-def gdf_conversion(gdf, name=None, return_type=None):
-    """can convert a GeoGataFrame(gdf) into GeoJSON or GPKG, but always returns the gdf and name"""
-    if return_type == "GeoJSON":
-        print(f"Writing down the geojson file {name + '.geojson'} ...")
-        start_time = time.time()
-        gdf.to_file(Path(__file__).parent/'data'/'output'/(name + '.geojson'), driver=return_type)
-        print(f"Writing file {time.time() - start_time} seconds ---")
-        print(f"GeoJSON {name + '.geojson'} was written.")
-        return gdf, name
-    elif return_type == "GPKG":
-        print(f"Writing down the geopackage file {name + '.gpkg'} ...")
-        start_time = time.time()
-        gdf.to_file(Path(__file__).parent/'data'/'output'/(name + '.gpkg'), driver=return_type)
-        print(f"Writing file {time.time() - start_time} seconds ---")
-        print(f"GeoPackage {name + '.gpkg'} was written.")
-        return gdf, name
-    else:
-        return gdf, name
+from other.utility_functions import gdf_conversion
 
 # Function for collection data from OSM dbf and conversion to GeoJSON
 # Could be extended with varios type of searches and filters
@@ -213,5 +99,3 @@ def osm_collect_buildings(name='buildings', driver=None):
     buildings = osm.get_buildings()
 
     return gdf_conversion(buildings, name, return_type=driver)
-
-# %%
